net: mqtt-sn: Rename internal 'find' functions
Add word 'by' to function names to clarify which is the thing that
is found and which is the thing that is searched for. For example,
mqtt_sn_topic_find_name was renamed to mqtt_sn_topic_find_by_name
because the function does not find the name but finds a topic by
its name.

Signed-off-by: Rene Beckmann <rene.bckmnn@gmail.com>